Refactor the main loop
Was experiencing index incrementing beyond for() limits at random times,
causing controller to get stuck infinite loop and keyboard to locked up.

Suspect non-volatile declaration for i.

https://www.arduino.cc/reference/en/language/variables/variable-scope-qualifiers/volatile/

Refactor to take out the body of the loop into the success path, or
return.

This refactor still has a high cyclomatic complexity, but as-is it
resolved the fault.